refactor SavingSummaryView Model
Related Work Items: #929

+++ GC.MFI.Solution/GC.MFI.Models/ViewModels/SavingSummaryViewModel.cs
@@ -66,7 +66,7 @@ namespace GC.MFI.Models.ViewModels
 
         public byte? SavingStatus { get; set; }
 
-        public short? EmployeeId { get; set; }
+        public long? EmployeeId { get; set; }
 
         public byte? MemberCategoryID { get; set; }
